html/semantics/forms/textfieldselection/selection-not-application-week-month.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>text field selection</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel="author" href="mailto:jarhar@chromium.org">
<link rel=help href="https://html.spec.whatwg.org/multipage/#textFieldSelection">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
  var nonApplicableTypes = ["month", "week"];

  nonApplicableTypes.forEach(function(type){
    var el = document.createElement("input");
    el.type = type;

    test(() => {
      assert_equals(el.selectionStart, null);
    }, `selectionStart on an input[type=${type}] returns null`);

    test(() => {
      assert_equals(el.selectionEnd, null);
    }, `selectionEnd on an input[type=${type}] returns null`);

    test(() => {
      assert_equals(el.selectionDirection, null);
    }, `selectionDirection on an input[type=${type}] returns null`);

    test(() => {
      assert_throws_dom("InvalidStateError", function(){
        el.selectionStart = 0;
      });
    }, `assigning selectionStart on an input[type=${type}] throws InvalidStateError`);

    test(() => {
      assert_throws_dom("InvalidStateError", function(){
        el.selectionEnd = 0;
      });
    }, `assigning selectionEnd on an input[type=${type}] throws InvalidStateError`);

    test(() => {
      assert_throws_dom("InvalidStateError", function(){
        el.selectionDirection = 'none';
      });
    }, `assigning selectionDirection on an input[type=${type}] throws InvalidStateError`);

    test(() => {
      assert_throws_dom("InvalidStateError", function(){
        el.setRangeText("foobar");
      });
    }, `setRangeText on an input[type=${type}] throws InvalidStateError`);

    test(() => {
      assert_throws_dom("InvalidStateError", function(){
        el.setSelectionRange(0, 1);
      });
    }, `setSelectionRange on an input[type=${type}] throws InvalidStateError`);
  });
</script>
